backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
pytz==2024.1

# ML/Data Analysis
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2

# PDF Generation & Visualization
reportlab==4.0.8         # latest version you already had
Pillow==10.0.1           # required by reportlab for image support
matplotlib==3.8.2        # you had a newer version, keep that
seaborn==0.13.0          # newer than 0.12.2

# File handling
aiofiles==23.2.1         # good for async file I/O
pathlib2==2.3.7          # for enhanced path handling (esp. for legacy support)

# Templating
jinja2==3.1.2
